--- spring-integration-file/src/main/java/org/springframework/integration/file/test/package-info.java
@@ -1,5 +1,0 @@
-/**
- * Provides classes used to support common test cases in modules
- * that are dependent on the file module.
- */
-package org.springframework.integration.file.test;
