+++ spring-integration-ftp/src/test/java/org/springframework/integration/ftp/outbound/BigMGetTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2002-2012 the original author or authors.
+ * Copyright 2002-2016 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -30,10 +30,11 @@ import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
  */
 @ContextConfiguration
 @RunWith(SpringJUnit4ClassRunner.class)
-public class BigMGetTests extends org.springframework.integration.file.test.BigMGetTests {
+public class BigMGetTests extends org.springframework.integration.file.BigMGetTests {
 
 	@Test @Ignore // needs directories and server (FTP and SFTP)
 	public void doTest() throws Exception {
 		assertEquals(FILES, this.mgetManyFiles().getPayload().size());
 	}
+
 }
